Move shapes to color_shapes
These are all using ColorVertex objects, so naming the file accordingly.
