+++ lib/graphics/src/color_shapes.rs
