+++ lib/graphics/src/lib.rs
@@ -5,16 +5,16 @@ mod app;
 mod program;
 mod shader;
 mod shader_source;
-mod shapes;
+mod color_shapes;
 mod texture;
 mod vertex;
 mod window;
 
 pub use app::App;
 pub use app::RenderingSource;
-pub use shapes::SimpleRect;
-pub use shapes::SimpleTriangle;
-pub use shapes::Updateable;
+pub use color_shapes::SimpleRect;
+pub use color_shapes::SimpleTriangle;
+pub use color_shapes::Updateable;
 pub use texture::TextureSetupDefinition;
 pub use vertex::VertexSpecable;
 pub use vertex::VertexSpecification;
